requirements.txt
fastapi~=0.121.0
uvicorn[standard]
sqlmodel~=0.0.27
qbittorrent-api
python-dotenv~=1.2.1
httpx
python-multipart
pydantic~=2.12.4
sqlalchemy~=2.0.44
requests~=2.32.5
bencodepy~=0.9.5
passlib~=1.7.4
PyJWT~=2.10.1